# backend/docs/swagger.yaml
basePath: /api
definitions:
  internal_server_handlers.CreateUserParams:
    properties:
      admin:
        type: boolean
      email:
        type: string
      lastLogin:
        $ref: "#/definitions/pgtype.Timestamptz"
      name:
        type: string
      password:
        type: string
      role:
        type: string
      status:
        type: string
    type: object
  internal_server_handlers.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
...
  internal_server_handlers.ClaimctlUser:
    properties:
      admin:
        type: boolean
      createdAt:
        $ref: "#/definitions/pgtype.Timestamptz"
      email:
        type: string
      id:
        type: integer
      lastLogin:
        $ref: "#/definitions/pgtype.Timestamptz"
      name:
        type: string
      password:
        type: string
      role:
        type: string
      slackDestination:
        $ref: "#/definitions/pgtype.Text"
      status:
        type: string
      teamsWebhookUrl:
        $ref: "#/definitions/pgtype.Text"
      updatedAt:
        $ref: "#/definitions/pgtype.Timestamptz"
    type: object
...
  pgtype.InfinityModifier:
    enum:
      - 1
      - 0
      - -1
    format: int32
    type: integer
    x-enum-varnames:
      - Infinity
      - Finite
      - NegativeInfinity
...
  pgtype.Text:
    properties:
      string:
        type: string
      valid:
        type: boolean
    type: object
  pgtype.Timestamptz:
    properties:
      infinityModifier:
        $ref: "#/definitions/pgtype.InfinityModifier"
      time:
        type: string
      valid:
        type: boolean
    type: object
...
host: localhost:3000
info:
  contact:
    email: support@github.com/thetaqitahmid/claimctl
    name: API Support
  description: API documentation for claimctl resource management system.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: claimctl API
  version: "1.0"
paths:
  /login:
    post:
      consumes:
        - application/json
      description: Authenticate a user and return a JWT token.
      parameters:
        - description: Login Credentials
          in: body
          name: loginRequest
          required: true
          schema:
            $ref: "#/definitions/internal_server_handlers.LoginRequest"
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: User Login
      tags:
        - auth
  /me:
    get:
      consumes:
        - application/json
      description: Retrieve details of the currently authenticated user.
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current user
      tags:
        - users
...
  /users:
    get:
      consumes:
        - application/json
      description: Retrieve a list of all users. Requires admin privileges.
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: "#/definitions/internal_server_handlers.ClaimctlUser"
            type: array
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get all users
      tags:
        - users
    post:
      consumes:
        - application/json
      description: Create a new user account. Requires admin privileges.
      parameters:
        - description: New User Details
          in: body
          name: user
          required: true
          schema:
            $ref: "#/definitions/internal_server_handlers.CreateUserParams"
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: "#/definitions/internal_server_handlers.ClaimctlUser"
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create a new user
      tags:
        - users
  /users/find:
    get:
      consumes:
        - application/json
      description: Retrieve a user by ID or email. Requires admin privileges.
      parameters:
        - description: User ID
          in: query
          name: id
          type: integer
        - description: User Email
          in: query
          name: email
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: "#/definitions/internal_server_handlers.ClaimctlUser"
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get a user
      tags:
        - users
swagger: "2.0"